+++ core/pfe-core/controllers/floating-dom-controller.ts
@@ -14,7 +14,7 @@ import {
   arrow as arrowMiddleware
 } from '@floating-ui/dom';
 
-type Lazy<T> = T|(() => T|null|undefined);
+type Lazy<T> = T | (() => T | null | undefined);
 
 interface FloatingDOMControllerOptions {
   content: Lazy<HTMLElement>;
@@ -32,8 +32,8 @@ interface ShowOptions {
   fallbackPlacements?: Placement[]
 }
 
-export type Anchor = ''|'top'|'left'|'bottom'|'right';
-export type Alignment = 'center'|'start'|'end';
+export type Anchor = '' | 'top' | 'left' | 'bottom' | 'right';
+export type Alignment = 'center' | 'start' | 'end';
 
 /**
  * Controls floating DOM within a web component, e.g. tooltips and popovers