+++ core/pfe-core/controllers/roving-tabindex-controller.ts
@@ -11,17 +11,22 @@ const isFocusableElement = (el: Element): el is HTMLElement =>
  * Components Using a Roving
  * tabindex](https://www.w3.org/WAI/ARIA/apg/practices/keyboard-interface/#kbd_roving_tabindex)
  */
-export class RovingTabindexController implements ReactiveController {
+export class RovingTabindexController<
+  ItemType extends HTMLElement = HTMLElement,
+> implements ReactiveController {
   /** active focusable element */
-  #activeItem?: HTMLElement;
+  #activeItem?: ItemType;
+
+  /** closest ancestor containing items */
+  #itemsContainer?: HTMLElement;
 
   /** array of all focusable elements */
-  #items: HTMLElement[] = [];
+  #items: ItemType[] = [];
 
   /**
    * finds focusable items from a group of items
    */
-  get #focusableItems(): HTMLElement[] {
+  get #focusableItems(): ItemType[] {
     return this.#items.filter(isFocusableElement);
   }
 
@@ -42,38 +47,38 @@ export class RovingTabindexController implements ReactiveController {
   /**
    * active item of array of items
    */
-  get activeItem(): HTMLElement | undefined {
+  get activeItem(): ItemType | undefined {
     return this.#activeItem;
   }
 
   /**
    * first item in array of focusable items
    */
-  get firstItem(): HTMLElement | undefined {
+  get firstItem(): ItemType | undefined {
     return this.#focusableItems[0];
   }
 
   /**
    * last item in array of focusable items
    */
-  get lastItem(): HTMLElement | undefined {
+  get lastItem(): ItemType | undefined {
     return this.#focusableItems.at(-1);
   }
 
   /**
    * next item  after active item in array of focusable items
    */
-  get nextItem(): HTMLElement | undefined {
+  get nextItem(): ItemType | undefined {
     return (
-        this.#activeIndex < this.#focusableItems.length - 1 ? this.#focusableItems[this.#activeIndex + 1]
-      : this.firstItem
+        this.#activeIndex >= this.#focusableItems.length - 1 ? this.firstItem
+      : this.#focusableItems[this.#activeIndex + 1]
     );
   }
 
   /**
    * previous item  after active item in array of focusable items
    */
-  get prevItem(): HTMLElement | undefined {
+  get prevItem(): ItemType | undefined {
     return (
         this.#activeIndex > 0 ? this.#focusableItems[this.#activeIndex - 1]
       : this.lastItem
@@ -87,7 +92,7 @@ export class RovingTabindexController implements ReactiveController {
   /**
    * handles keyboard navigation
    */
-  #onKeydown(event: KeyboardEvent):void {
+  #onKeydown = (event: KeyboardEvent) => {
     if (event.ctrlKey || event.altKey || event.metaKey || this.#focusableItems.length < 1) {
       return;
     }
@@ -97,8 +102,8 @@ export class RovingTabindexController implements ReactiveController {
     const horizontalOnly =
         !item ? false
       : item.tagName === 'SELECT' ||
-        item.getAttribute('aria-expanded') === 'true' ||
         item.getAttribute('role') === 'spinbutton';
+
 
     switch (event.key) {
       case 'ArrowLeft':
@@ -153,12 +158,12 @@ export class RovingTabindexController implements ReactiveController {
       event.stopPropagation();
       event.preventDefault();
     }
-  }
+  };
 
   /**
    * sets tabindex of item based on whether or not it is active
    */
-  updateActiveItem(item?: HTMLElement):void {
+  updateActiveItem(item?: ItemType): void {
     if (item) {
       if (!!this.#activeItem && item !== this.#activeItem) {
         this.#activeItem.tabIndex = -1;
@@ -171,15 +176,16 @@ export class RovingTabindexController implements ReactiveController {
   /**
    * focuses on an item and sets it as active
    */
-  focusOnItem(item?: HTMLElement):void {
+  focusOnItem(item?: ItemType): void {
     this.updateActiveItem(item || this.firstItem);
     this.#activeItem?.focus();
+    this.host.requestUpdate();
   }
 
   /**
    * Focuses next focusable item
    */
-  updateItems(items: HTMLElement[]) {
+  updateItems(items: ItemType[]) {
     const sequence = [...items.slice(this.#itemIndex), ...items.slice(0, this.#itemIndex)];
     const first = sequence.find(item => this.#focusableItems.includes(item));
     this.focusOnItem(first || this.firstItem);
@@ -188,7 +194,7 @@ export class RovingTabindexController implements ReactiveController {
   /**
    * from array of HTML items, and sets active items
    */
-  initItems(items: HTMLElement[]) {
+  initItems(items: ItemType[], itemsContainer: HTMLElement = this.host) {
     this.#items = items ?? [];
     const focusableItems = this.#focusableItems;
     const [focusableItem] = focusableItems;
@@ -196,9 +202,27 @@ export class RovingTabindexController implements ReactiveController {
     for (const item of focusableItems) {
       item.tabIndex = this.#activeItem === item ? 0 : -1;
     }
+    /**
+     * removes listener on previous contained and applies it to new container
+     */
+    if (!this.#itemsContainer || itemsContainer !== this.#itemsContainer) {
+      this.#itemsContainer?.removeEventListener('keydown', this.#onKeydown);
+      this.#itemsContainer = itemsContainer;
+      this.hostConnected();
+    }
   }
 
+  /**
+   * adds event listeners to items container
+   */
   hostConnected() {
-    this.host.addEventListener('keydown', this.#onKeydown.bind(this));
+    this.#itemsContainer?.addEventListener('keydown', this.#onKeydown);
+  }
+
+  /**
+   * removes event listeners from items container
+   */
+  hostDisconnected() {
+    this.#itemsContainer?.removeEventListener('keydown', this.#onKeydown);
   }
 }