+++ docs/_plugins/pfe-assets.cjs
@@ -72,6 +72,7 @@ async function bundle() {
 
 module.exports = {
   configFunction(eleventyConfig, options) {
+    eleventyConfig.addPassthroughCopy('docs/images/favicon.ico');
     eleventyConfig.addPassthroughCopy('docs/bundle.{js,map,ts}');
     eleventyConfig.addPassthroughCopy('docs/pfe.min.{map,css}');
     eleventyConfig.addPassthroughCopy({ 'elements/pfe.min.*': '/' } );
@@ -96,5 +97,3 @@ module.exports = {
     eleventyConfig.on('eleventy.before', () => bundle(options));
   },
 };
-
-