+++ elements/pf-jump-links/test/pf-jump-links.spec.ts
@@ -22,29 +22,30 @@ async function allUpdates(element: ReactiveElement) {
 
 describe('<pf-jump-links>', function() {
   let element: PfJumpLinks;
+  let firstItem: PfJumpLinksItem;
+  let secondItem: PfJumpLinksItem;
 
   beforeEach(async function() {
     element = await createFixture<PfJumpLinks>(html`
       <pf-jump-links>
         <pf-jump-links-item id="first">Inactive section</pf-jump-links-item>
-        <pf-jump-links-item id="second">Active section</pf-jump-links-item>
+        <pf-jump-links-item id="second" active>Active section</pf-jump-links-item>
         <pf-jump-links-item id="third">Inactive section</pf-jump-links-item>
       </pf-jump-links>
     `);
     await allUpdates(element);
+    [firstItem, secondItem] = element.querySelectorAll<PfJumpLinksItem>('pf-jump-links-item');
   });
 
   describe('tabbing to first item', function() {
-    let firstItem: PfJumpLinksItem;
-    let secondItem: PfJumpLinksItem;
-    let initialActiveElement: Element|null;
+    let initialActiveElement: Element | null;
     beforeEach(async function() {
-      [firstItem, secondItem] = element.querySelectorAll<PfJumpLinksItem>('pf-jump-links-item');
-      await sendKeys({ press: 'Tab' });
       initialActiveElement = document.activeElement;
+      await sendKeys({ press: 'Tab' });
+      await nextFrame();
     });
 
-    it('should focus a this first jump-links-item', function() {
+    it('should focus the first jump-links-item', function() {
       expect(document.activeElement).to.equal(firstItem);
     });