+++ elements/pf-switch/pf-switch.ts
@@ -8,6 +8,8 @@ import styles from './pf-switch.css';
  * checkboxes can often be used interchangeably, but the switch provides a more
  * explicit, visible representation on a setting.
  *
+ * @fires {Event} change - Fires when the switch selection changes.
+ *
  * @cssprop --pf-c-switch--FontSize {@default `1rem`}
  * @cssprop {<length>} --pf-c-switch--ColumnGap {@default `1rem`}
  * @cssprop --pf-c-switch__toggle-icon--FontSize {@default `calc(1rem * .625)`}